+++ build.h
@@ -1,11 +1,11 @@
-#define BUILD_NUMBER 26
+#define BUILD_NUMBER 27
 #define MEJOR_VERSION 0
 #define MINOR_VERSION 1
 #define PATCH_NUMBER 0
-#define VERSION_STRING "0.1.26.0\0"
-#define VERSION_STR_DWORD 0,1,26,0
+#define VERSION_STRING "0.1.27.0\0"
+#define VERSION_STR_DWORD 0,1,27,0
 #define PRODUCT_VER_STRING "0.1\0"
 #define PRODUCT_VER_DWORD 0,1,0,0
 #define COMMON_PRODUCTNAME "VolumeWalker\0"
-#define COMMON_LEGALCOPYRIGHT "Copyright (C) 2023-2024 YAMASHITA Katsuhiro"
+#define COMMON_LEGALCOPYRIGHT "Copyright (C) 2023-2025 YAMASHITA Katsuhiro"
 #define COMMON_COMPANYNAME "YAMASHITA Katsuhiro, Yamashita Software Works"